# src/main/resources/openapi/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/password/request:
    post:
      tags:
      - password-recovery-controller
      operationId: request
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestDTO"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "200":
          description: OK
  /api/password/confirm:
    post:
      tags:
      - password-recovery-controller
      operationId: confirm
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmResetDTO"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "200":
          description: OK
  /api/password/validate:
    get:
      tags:
      - password-recovery-controller
      operationId: validate
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenValidationResponseDTO"
components:
  schemas:
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        path:
          type: string
    PasswordResetRequestDTO:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
      required:
      - email
    ConfirmResetDTO:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
      - token
    TokenValidationResponseDTO:
      type: object
      properties:
        valid:
          type: boolean